Fix variant load hangup
Fixed the tendency for a hangup when attempting to load a gene's
variants. There's only a small subset of fields that need to be
displayed, but prior to the fix all of the fields for a variant
were being queried. This is still quite slow when many variants need
to be loaded, but better than before. Hangup will presumably still
occur if the variant list gets large enough.